Change types (due to compilation error on windows)

+++ src/misc/impl/ObsfucatedString.hpp
@@ -36,7 +36,7 @@ class ObsfucatedString
 {
 public:
 
-    typedef u_int64_t LONG;
+    typedef unsigned long long LONG;
 
     /** Constructor. */
     ObsfucatedString (LONG l, ...);
